+++ src/eval/metrics.h
@@ -13,7 +13,6 @@ float mse(const VectorXf& y, const VectorXf& yhat, VectorXf& loss,
             const vector<float>& class_weights=vector<float>() );
 
 // TODO: test cases for the metrics
-// TODO: implement the metrics for classification
 
 /// log loss (2 methods below)
 VectorXf log_loss(const VectorXf& y, const VectorXf& predict_proba,
